Improve comments; confirm deletion of database

+++ database/dbm.py
@@ -23,6 +23,7 @@ class DatabaseManager:
         self.con = sqlite3.connect(filename, detect_types=sqlite3.PARSE_DECLTYPES)
         self.cur = self.con.cursor()
 
+        # Create the table if it doesn't already exist. Note the columns are defined here
         self.cur.execute('''CREATE TABLE IF NOT EXISTS activities
             (
              `job` VARCHAR(100) DEFAULT '',  
@@ -37,7 +38,8 @@ class DatabaseManager:
         self.con.close()
 
     def add_activity(self, desc: str, start: datetime.datetime, duration: datetime.timedelta) -> int:
-        '''Creates an activity given a description, start time, and duration.
+        '''Creates an activity given a description, start time, and duration. Returns the row
+        id of that insert if successful, -1 if not.
 
         Sample usage:
         >>> x = DatabaseManger("CIS 211 LA")
@@ -46,32 +48,51 @@ class DatabaseManager:
         >>> self.add_activitiy("Office hours", start, duration)
         '''
         try: 
+            # Attempt to insert the requested activity into the database
             self.cur.execute('''INSERT INTO activities
                 (job, desc, start_time, duration)
                 values
                 (?, ?, ?, ?)
             ''', (self.job, desc, start, duration.total_seconds()))
 
+            # Commit that statement to the database
             self.con.commit()
+
+            # We return the id of the row that we just created
             return self.cur.lastrowid
         except:
+            # If we fail somewhere in the above try statement, we must rollback (reset) the
+            # transaction and return -1, meaning that we failed.
             self.con.rollback()
             return -1
-    def delete_activity(self, id: int) -> int:
+    def delete_activity(self, id: int) -> bool:
+        '''Deletes an actvity given an ID. Returns true if successful and false if not'''
         try:
+            # Attempt to delete from the database where the rowid matches
             self.cur.execute('DELETE FROM activities where (rowid = ?)',
                     (id,)
             )
             self.con.commit()
+
+            # If there was a row that was changed (deleted) we return true: successful
             if self.cur.rowcount == 1:
                 return True
             else:
+                # Otherwise, we didn't delete anything, return false
                 return False
         except:
+            # We also return false if there was an error
+            # TODO: We may want to do some further error handling here. Either
+            # printing that there is an error or re-init'ing the database
             return False
 
 
     def debug_get_all(self) -> list:
+        '''Returns every single element from the database with a rowid. See 
+        get_activities_within_range() docstring for details on formatting
+
+        Not supposed to be used in production. Very helpful for testing.
+        '''
         self.cur.execute('SELECT rowid,* FROM activities')
         return self.cur.fetchall()
 
@@ -81,28 +102,34 @@ class DatabaseManager:
             rowid:int, job:str, description:str, start:datetime.datetime, duration:int
         )
 
-
         Sample usage:
         >>> x = DatabaseManager("CIS 211 LA")
         >>> now = datetime.datetime.now()
         >>> # Gets all activities within the last 3 hours
         >>> x.get_activies_within_range(now - datetime.timedelta(hours=3), now)
         '''
+
+        # We execute a select statement that selects activities that _start_ in the
+        # specified range and then return what it yields
         self.cur.execute('''
-                SELECT rowid,* FROM activities where not (start_time < ? or ? < start_time)''', 
-                (start, end)
+                SELECT rowid,* FROM activities where not (start_time < ? or ? < start_time)
+                and (job=?)''', 
+                (start, end, self.job)
         )
 
         return self.cur.fetchall()
     
-    def get_activity(self, rowid: int):
+    def get_activity(self, rowid: int) -> tuple:
         '''
         Returns a single activity as a tuple of (
             rowid:int, job:str, description:str, start:datetime.datetime, duration:int
         )
 
+        Returns None if the object doesn't exist. 
         '''
 
+        # We select a row based on the rowid and return what that statement yields
+        # NOTE: This is not bound by self.job.
         self.cur.execute('Select rowid,* FROM activities where rowid=?',
                 (rowid,)
         )
@@ -118,30 +145,53 @@ class DatabaseManager:
         >>> edit_activity(3, duration=datetime.timedelta(hours=1.10))
         '''
 
+        # Keep track of how many rows get updated
         max_rowcount = 0
 
+        # Convert duration to seconds or None (if unspecified) for the following line
         dur = None if duration is None else duration.seconds
+
+        # Store all of the fields in tuples, formatted as (field_name, value). Note that
+        # field_name must exactly match what it is in the SQL database
         fields = [('job', job), ('desc', description), ('start_time', start), ('duration', dur)]
 
+        # we loop through each field
         for field in fields:
             if field[1] is None:
+                # Skipping the ones where the value is None
                 continue
-            self.cur.execute(f'UPDATE activities SET {field[0]} = ? WHERE rowid = ?', (field[1], rowid))
+            # And update the field of that rowid with the value
+            self.cur.execute(f'UPDATE activities SET {field[0]} = ? WHERE rowid = ?', 
+                    (field[1], rowid))
             self.con.commit()
+
+            # And, whenever we make a change, we track the total number of rows modified.
+            # Remember that the rowid is the same each time, so we want to track if the
+            # number of rows modified is 0 or 1.
             max_rowcount = max(max_rowcount, self.cur.rowcount)
 
+        # if in our tracking of modified rows we modified a row, we return True, otherwise False
         if max_rowcount == 1:
             return True
         else:
             return False
 
 
-    def reset_db(self):
-        self.cur.execute('''
-            drop table activities;
-        ''')
-        self.con.commit()
-        self.__init__(self.job, self.filename)
+    def reset_db(self, sure: str=''):
+
+        # If we got a sure response, we delete everything in the database and re-create it
+        if sure == 'yes':
+            self.cur.execute('''
+                drop table activities;
+            ''')
+            self.con.commit()
+            self.__init__(self.job, self.filename)
+        else:
+            print(sure)
+            # Otherwise, print out some warning and instructions to delete everything
+            print("Are you sure you'd like to delete this database and everything")
+            print("in it? If so, call this function with 'yes' as an argument")
+            print("example: db.reset_db('yes')")
